DEV : add validations to user input

+++ src/main/java/org/cdl/object/BookingItem.java
@@ -30,7 +30,7 @@ public class BookingItem {
      * recalculate price based on unit price and discount price schemes
      */
     public void recalculateItemPrice() {
-        int tempPrice = 0;
+        double tempPrice = 0;
         int balanceQuantity = quantity;
         for (PriceScheme scheme : priceSchemes) {
             if (balanceQuantity >= scheme.getQuantity()) {
